+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/KubernetesClient.java
@@ -66,12 +66,16 @@ import io.fabric8.kubernetes.api.model.DoneableService;
 import io.fabric8.kubernetes.api.model.ServiceAccount;
 import io.fabric8.kubernetes.api.model.ServiceAccountList;
 import io.fabric8.kubernetes.api.model.DoneableServiceAccount;
+import io.fabric8.kubernetes.api.model.coordination.v1.DoneableLease;
+import io.fabric8.kubernetes.api.model.coordination.v1.Lease;
+import io.fabric8.kubernetes.api.model.coordination.v1.LeaseList;
 import io.fabric8.kubernetes.client.dsl.*;
 import io.fabric8.kubernetes.api.model.apiextensions.CustomResourceDefinition;
 import io.fabric8.kubernetes.api.model.apiextensions.CustomResourceDefinitionList;
 import io.fabric8.kubernetes.api.model.apiextensions.DoneableCustomResourceDefinition;
 import io.fabric8.kubernetes.client.dsl.base.CustomResourceDefinitionContext;
 import io.fabric8.kubernetes.client.dsl.internal.RawCustomResourceOperationsImpl;
+import io.fabric8.kubernetes.client.extended.leaderelection.LeaderElectorBuilder;
 import io.fabric8.kubernetes.client.informers.SharedInformerFactory;
 
 import java.io.InputStream;
@@ -420,4 +424,18 @@ public interface KubernetesClient extends Client {
    */
   SharedInformerFactory informers(ExecutorService executorService);
 
+  /**
+   * API entrypoint for <code>LeaderElector</code> implementation for leader election.
+   *
+   * @param <C> type parameter for the Namespaceable KubernetesClient
+   * @return LeaderElectorBuilder to build LeaderElector instances
+   */
+  <C extends Namespaceable<C> & KubernetesClient> LeaderElectorBuilder<C> leaderElector();
+
+  /**
+   * API entrypoint for {@link Lease} related operations. Lease (coordination.k8s.io/v1)
+   *
+   * @return MixedOperation object for Lease related operations.
+   */
+  MixedOperation<Lease, LeaseList, DoneableLease, Resource<Lease, DoneableLease>> leases();
 }