+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Update Jackson Bom to 2.10.2
 
 #### New Features
+* Fix #1948: LeaderElection Implementation (Lock Implementations: ConfigMap & Lease)
 
 
 ### 4.7.1 (24-01-2020)